allow float starting balance

+++ cmd.py
@@ -12,7 +12,7 @@ def main(params):
     budget_simulator = BudgetSimulator(config = params['config'],
                                         start_date = params.get('start_date'),
                                         end_date = params.get('end_date'),
-                                        start_balance = int(params.get('balance')))
+                                        start_balance = float(params.get('balance')))
 
 
     if params.get('to_csv') == "True":
